[enh] global strategy to keep old events

+++ events/management/commands/fetch_events.py
@@ -52,16 +52,23 @@ class Command(BaseCommand):
                 print e
 
 
-def event_source(background_color, text_color):
+def event_source(background_color, text_color, key="url"):
     def event_source_wrapper(func, org_name=None):
         def fetch_events(quiet):
             def create_event(**detail):
+                if key not in (None, False) and Event.objects.filter(**{key: detail[key]}):
+                    Event.objects.filter(**{key: detail[key]}).delete()
+
                 res = Event.objects.create(source=org_name, text_color=COLORS[org_name]["fg"], border_color=COLORS[org_name]["bg"], **detail)
                 if not quiet:
                     print "[%s] %s (%s)" % (res.source, res.title, res.start)
                 return res
 
-            Event.objects.filter(source=org_name).delete()
+            if key in (None, False):
+                Event.objects.filter(source=org_name).delete()
+            else:
+                Event.objects.filter(source=org_name, start__gte=datetime.now())
+
             func(create_event)
             if not quiet:
                 print " === Finished for " + org_name
@@ -94,7 +101,7 @@ def json_api(org_name, url, background_color, text_color):
                 location=event['location'] if 'location' in event else None,
             )
 
-    return event_source(background_color, text_color)(fetch, org_name)
+    return event_source(background_color, text_color, key=None)(fetch, org_name)
 
 
 def generic_meetup(org_name, meetup_name, background_color, text_color):
@@ -124,7 +131,7 @@ def generic_meetup(org_name, meetup_name, background_color, text_color):
     return event_source(background_color, text_color)(fetch, org_name)
 
 
-@event_source(background_color="#133F52", text_color="#FFFFFF")
+@event_source(background_color="#133F52", text_color="#FFFFFF", key=None)
 def afpyro(create_event):
     soup = BeautifulSoup(urlopen("http://afpyro.afpy.org/").read())
     filtering = lambda x: x['href'][:7] == '/dates/' and '(BE)' in x.text
@@ -191,7 +198,7 @@ def bhackspace(create_event):
 generic_meetup("bigdata_be", "bigdatabe", background_color="black", text_color="white")
 
 
-@event_source(background_color="#828282", text_color="white")
+@event_source(background_color="#828282", text_color="white", key=None)
 def blender_brussels(create_event):
     soup = BeautifulSoup(urlopen("https://blender-brussels.github.io/"))
 
@@ -255,7 +262,7 @@ def bxlug(create_event):
         )
 
 
-@event_source(background_color="#D2C7BA", text_color="black")
+@event_source(background_color="#D2C7BA", text_color="black", key=None)
 def constantvzw(create_event):
     soup = BeautifulSoup(urlopen("http://www.constantvzw.org/site/").read())
 
@@ -299,7 +306,7 @@ generic_meetup("docker_belgium", "Docker-Belgium", background_color="#008FC4", t
 generic_meetup("ember_js_brussels", "Ember-js-Brussels", background_color="#FC745D", text_color="white")
 
 
-@event_source(background_color="#C9C4BF", text_color="black")
+@event_source(background_color="#C9C4BF", text_color="black", key=None)
 def foam(create_event):
     soup = BeautifulSoup(urlopen("http://fo.am/events/").read())
 
@@ -325,7 +332,7 @@ def foam(create_event):
             event.tags.add("meeting")
 
 
-@event_source(background_color="coral", text_color="white")
+@event_source(background_color="coral", text_color="white", key=None)
 def hsbxl(create_event):
     today = date.today() - timedelta(days=6 * 30)
 
@@ -399,7 +406,7 @@ def neutrinet(create_event):
             event.tags.add("meeting")
 
 
-@event_source(background_color="#299C8F", text_color="white")
+@event_source(background_color="#299C8F", text_color="white", key=None)
 def okfnbe(create_event):
     data = Calendar.from_ical(urlopen("https://www.google.com/calendar/ical/sv07fu4vrit3l8nb0jlo8v7n80@group.calendar.google.com/public/basic.ics").read())
 
@@ -470,7 +477,7 @@ def owaspbe(create_event):
 generic_meetup("phpbenelux", "phpbenelux", background_color="#015074", text_color="white")
 
 
-@event_source(background_color="#2BC884", text_color="white")
+@event_source(background_color="#2BC884", text_color="white", key=None)
 def relab(create_event):
     data = Calendar.from_ical(urlopen("https://www.google.com/calendar/ical/utmnk71g19dcs2d0f88q3hf528%40group.calendar.google.com/public/basic.ics").read())
 
@@ -504,7 +511,7 @@ generic_meetup("ruby_burgers", "ruby_burgers-rb", background_color="white", text
 json_api("urlab", "https://urlab.be/hackeragenda.json", background_color="pink", text_color="black")
 
 
-@event_source(background_color="#25272C", text_color="#C58723")
+@event_source(background_color="#25272C", text_color="#C58723", key=None)
 def voidwarranties(create_event):
     data = Calendar.from_ical(urlopen("http://voidwarranties.be/index.php/Special:Ask/-5B-5BCategory:Events-5D-5D/-3FHas-20start-20date=start/-3FHas-20end-20date=end/-3FHas-20coordinates=location/format=icalendar/title=VoidWarranties/description=Events-20at-20http:-2F-2Fvoidwarranties.be/limit=500").read())
